Add support for the Mocha API.
Mocha throws an error when it runs `exports.xit = context.xit ||
context.test.skip` because mocha-cakes-2 has not set `context.xit` or
`context.test`.

When Mocha is run using the CLI the default ui is set up first, which
sets `context.xit`, and then that setup is overriden by the ui you've
specified in the options, so the non-default ui can "inherit" parts of
the default ui by not overriding them.

But when you use the API and construct Mocha with the ui option (`const
mocha = new Mocha({ui: 'mocha-cakes-2'})`) the default ui isn't set up,
so Mocha runs into that error.